fix (sample_sent, sample_labeled_sent): change logic to determine the SENT sequence length
When dealing with trees, which have a low amount of edges compared to the number of nodes, the `seq_length` of the SENT can be underestimated. This leads to various buffer overflows. Here, we derive the length of a SENT based on the space complexity described in the paper of n_edges*n_nodes. We scale it appropriately for node features and edge features.